Moving open_file to .utils.basics; adding docstrings
