+++ art/__init__.py
@@ -4,6 +4,8 @@
 Academic Review Tool (ART)
 ==========================
 
+Version: 1.1.0
+
 The Academic Review Tool (ART) is a package for performing academic reviews and bibliometric analyses in Python. 
 It offers capabilities for discovering, retrieving, and analysing academic literature at scale. 
 ART accesses records from Crossref, Web of Science, Scopus, Orcid, and more.
@@ -28,10 +30,9 @@ ART uses the following APIs:
 * Scopus
 * ORCID
 * Geopy / Nominatim
-
-Version: 1.1.0
 """
 
+from .utils.basics import open_file as open
 from .importers.crossref import lookup_doi, lookup_dois, lookup_journal, lookup_journals, search_journals, get_journal_entries, search_journal_entries, lookup_funder, lookup_funders, search_funders, get_funder_works, search_funder_works
 from .importers.crossref import search_works as search_crossref
 # from .importers.wos import search as search_wos
@@ -42,16 +43,3 @@ from .importers.search import search as api_search
 from .classes import Results, References, Author, Authors, Funder, Funders, Affiliation, Affiliations, Review
 from .classes.networks import Network, Networks
 from .classes.citation_crawler import academic_scraper as scrape
-
-import pickle
-
-def open_file(file_address: str = 'request_input'): # type: ignore
-
-        if file_address == 'request_input':
-            file_address = input('File address: ')
-        
-        if (file_address.endswith('.txt')) or (file_address.endswith('.review')):
-            with open(file_address, 'rb') as f:
-                review = pickle.load(f)
-
-        return review
