+++ art/utils/basics.py
@@ -5,6 +5,7 @@ import math
 from os import stat_result
 import os
 import sys
+import pickle
 
 results_cols = [
                             'work_id',
@@ -267,5 +268,20 @@ def stat_file_to_dict(self):
         pass    
     
     return dictionary
+
+def open_file(file_address: str = 'request_input'): # type: ignore
+
+        """
+        Reads saved ART files and returns as a Review object. Reviews must be formatted as .review or pickled .txt files.
+        """
+
+        if file_address == 'request_input':
+            file_address = input('File address: ')
+        
+        if (file_address.endswith('.txt')) or (file_address.endswith('.review')):
+            with open(file_address, 'rb') as f:
+                review = pickle.load(f)
+
+        return review
 
 stat_result.to_dict = stat_file_to_dict # type: ignore
